Restructure validator_manager contracts and tests
Done:
        - Rename `contract_utils.py` to `validator_manager_utils.py`
        - Move `validator_manager_tester` to `test_validator_manager.py`
        - Move tests for `validator_manager_utils` to
          `test_validator_manager_utils.py`
TODO:
        - Add test cases and perform refactoring
        - Add `txToShard` and the remaining parts
