+++ sharding/validator_manager_utils.py
@@ -17,10 +17,39 @@ _valmgr_addr = None
 _valmgr_sender_addr = None
 _valmgr_tx = None
 
+viper_rlp_decoder_tx = rlp.decode(utils.parse_as_bin("0xf90237808506fc23ac00830330888080b902246102128061000e60003961022056600060007f010000000000000000000000000000000000000000000000000000000000000060003504600060c082121515585760f882121561004d5760bf820336141558576001905061006e565b600181013560f783036020035260005160f6830301361415585760f6820390505b5b368112156101c2577f010000000000000000000000000000000000000000000000000000000000000081350483602086026040015260018501945060808112156100d55760018461044001526001828561046001376001820191506021840193506101bc565b60b881121561014357608081038461044001526080810360018301856104600137608181141561012e5760807f010000000000000000000000000000000000000000000000000000000000000060018401350412151558575b607f81038201915060608103840193506101bb565b60c08112156101b857600182013560b782036020035260005160388112157f010000000000000000000000000000000000000000000000000000000000000060018501350402155857808561044001528060b6838501038661046001378060b6830301830192506020810185019450506101ba565bfe5b5b5b5061006f565b601f841315155857602060208502016020810391505b6000821215156101fc578082604001510182826104400301526020820391506101d8565b808401610420528381018161044003f350505050505b6000f31b2d4f"), Transaction)
+viper_rlp_decoder_addr = viper_rlp_decoder_tx.creates
+
+sighasher_tx = rlp.decode(utils.parse_as_bin("0xf9016d808506fc23ac0083026a508080b9015a6101488061000e6000396101565660007f01000000000000000000000000000000000000000000000000000000000000006000350460f8811215610038576001915061003f565b60f6810391505b508060005b368312156100c8577f01000000000000000000000000000000000000000000000000000000000000008335048391506080811215610087576001840193506100c2565b60b881121561009d57607f8103840193506100c1565b60c08112156100c05760b68103600185013560b783036020035260005101840193505b5b5b50610044565b81810360388112156100f4578060c00160005380836001378060010160002060e052602060e0f3610143565b61010081121561010557600161011b565b6201000081121561011757600261011a565b60035b5b8160005280601f038160f701815382856020378282600101018120610140526020610140f350505b505050505b6000f31b2d4f"), Transaction)
+sighasher_addr = sighasher_tx.creates
 
 class MessageFailed(Exception):
 
     pass
+
+
+def mk_validation_code(address):
+    '''
+    validation_code = """
+~calldatacopy(0, 0, 128)
+~call(3000, 1, 0, 0, 128, 0, 32)
+return(~mload(0) == {})
+    """.format(utils.checksum_encode(address))
+    return serpent.compile(validation_code)
+    '''
+    # The precompiled bytecode of the validation code which
+    # verifies EC signatures
+    validation_code_bytecode = b"a\x009\x80a\x00\x0e`\x009a\x00GV`\x80`\x00`\x007` "
+    validation_code_bytecode += b"`\x00`\x80`\x00`\x00`\x01a\x0b\xb8\xf1Ps"
+    validation_code_bytecode += address
+    validation_code_bytecode += b"`\x00Q\x14` R` ` \xf3[`\x00\xf3"
+    return validation_code_bytecode
+
+
+def sign(msg_hash, privkey):
+    v, r, s = utils.ecsign(msg_hash, privkey)
+    signature = utils.encode_int32(v) + utils.encode_int32(r) + utils.encode_int32(s)
+    return signature
 
 
 def get_valmgr_ct():
@@ -153,11 +182,6 @@ def call_validation_code(state, validation_code_addr, msg_hash, signature):
         raise MessageFailed()
     return bool(utils.big_endian_to_int(result))
 
-viper_rlp_decoder_tx = rlp.decode(utils.parse_as_bin("0xf90237808506fc23ac00830330888080b902246102128061000e60003961022056600060007f010000000000000000000000000000000000000000000000000000000000000060003504600060c082121515585760f882121561004d5760bf820336141558576001905061006e565b600181013560f783036020035260005160f6830301361415585760f6820390505b5b368112156101c2577f010000000000000000000000000000000000000000000000000000000000000081350483602086026040015260018501945060808112156100d55760018461044001526001828561046001376001820191506021840193506101bc565b60b881121561014357608081038461044001526080810360018301856104600137608181141561012e5760807f010000000000000000000000000000000000000000000000000000000000000060018401350412151558575b607f81038201915060608103840193506101bb565b60c08112156101b857600182013560b782036020035260005160388112157f010000000000000000000000000000000000000000000000000000000000000060018501350402155857808561044001528060b6838501038661046001378060b6830301830192506020810185019450506101ba565bfe5b5b5b5061006f565b601f841315155857602060208502016020810391505b6000821215156101fc578082604001510182826104400301526020820391506101d8565b808401610420528381018161044003f350505050505b6000f31b2d4f"), Transaction)
-viper_rlp_decoder_addr = viper_rlp_decoder_tx.creates
-
-sighasher_tx = rlp.decode(utils.parse_as_bin("0xf9016d808506fc23ac0083026a508080b9015a6101488061000e6000396101565660007f01000000000000000000000000000000000000000000000000000000000000006000350460f8811215610038576001915061003f565b60f6810391505b508060005b368312156100c8577f01000000000000000000000000000000000000000000000000000000000000008335048391506080811215610087576001840193506100c2565b60b881121561009d57607f8103840193506100c1565b60c08112156100c05760b68103600185013560b783036020035260005101840193505b5b5b50610044565b81810360388112156100f4578060c00160005380836001378060010160002060e052602060e0f3610143565b61010081121561010557600161011b565b6201000081121561011757600261011a565b60035b5b8160005280601f038160f701815382856020378282600101018120610140526020610140f350505b505050505b6000f31b2d4f"), Transaction)
-sighasher_addr = sighasher_tx.creates
 
 def mk_initiating_contracts(sender_privkey, sender_starting_nonce):
     o = []
@@ -170,81 +194,3 @@ def mk_initiating_contracts(sender_privkey, sender_starting_nonce):
         o.append(tx)
         nonce += 1
     return o
-
-
-# Testing Part
-def deploy_tx(state, tx):
-    success, output = apply_transaction(state, tx)
-    if not success:
-        raise t.TransactionFailed("Failed to deploy tx")
-    return output
-
-
-def deploy_contract(state, sender_privkey, bytecode):
-    tx = Transaction(
-            state.get_nonce(utils.privtoaddr(sender_privkey)),
-            GASPRICE, STARTGAS, to=b'', value=0,
-            data=bytecode
-    ).sign(sender_privkey)
-    return deploy_tx(state, tx)
-
-
-def deploy_initializing_contracts(sender_privkey, state):
-    sender_addr = utils.privtoaddr(sender_privkey)
-    txs = mk_initiating_contracts(t.k0, state.get_nonce(sender_addr))
-    for tx in txs:
-        try:
-            deploy_tx(state, tx)
-        except t.TransactionFailed:
-            pass
-
-
-def mk_validation_code(address):
-    '''
-    validation_code = """
-~calldatacopy(0, 0, 128)
-~call(3000, 1, 0, 0, 128, 0, 32)
-return(~mload(0) == {})
-    """.format(utils.checksum_encode(address))
-    return serpent.compile(validation_code)
-    '''
-    # The precompiled bytecode of the validation code which
-    # verifies EC signatures
-    validation_code_bytecode = b"a\x009\x80a\x00\x0e`\x009a\x00GV`\x80`\x00`\x007` "
-    validation_code_bytecode += b"`\x00`\x80`\x00`\x00`\x01a\x0b\xb8\xf1Ps"
-    validation_code_bytecode += address
-    validation_code_bytecode += b"`\x00Q\x14` R` ` \xf3[`\x00\xf3"
-    return validation_code_bytecode
-
-
-def sign(msg_hash, privkey):
-    v, r, s = utils.ecsign(msg_hash, privkey)
-    signature = utils.encode_int32(v) + utils.encode_int32(r) + utils.encode_int32(s)
-    return signature
-
-
-def test():
-    deposit_size = 10 ** 20
-    withdraw_hash = utils.sha3("withdraw")
-    valmgr_sender_privkey = t.k0
-    c = t.Chain()
-    c.mine(1, coinbase=t.a0)
-    state = c.head_state
-    state.gas_limit = 10 ** 10
-    state.set_balance(address=t.a0, value=deposit_size * 10)
-    state.set_balance(address=t.a1, value=deposit_size * 10)
-
-    deploy_initializing_contracts(t.k0, state)
-    validator_manager_addr = get_valmgr_addr()
-    k0_valcode_addr = deploy_contract(state, t.k0, mk_validation_code(t.a0))
-    tx = call_deposit(state, validator_manager_addr, t.k0, deposit_size, k0_valcode_addr, t.a2)
-    deploy_tx(state, tx)
-    assert hex(utils.big_endian_to_int(k0_valcode_addr)) == hex(utils.big_endian_to_int(call_sample(state, validator_manager_addr, 0)))
-    tx = call_withdraw(state, validator_manager_addr, t.k0, 0, sign(withdraw_hash, t.k0))
-    deploy_tx(state, tx)
-    assert 0 == utils.big_endian_to_int(call_sample(state, validator_manager_addr, 0))
-    assert call_validation_code(state, k0_valcode_addr, withdraw_hash, sign(withdraw_hash, t.k0))
-
-
-if __name__ == '__main__':
-    test()
